refactor(raft): move methods for reconfiguring out of RaftContext
This also moves `anoint` eventhough technicaly it is not a reconfiguring
operation. However, it is still something that is triggered external and
could fit in with the other operations.
